Part1 checked for project validation

+++ 42.1/Libft/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: jpirsch <jpirsch@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/11/06 18:07:18 by jpirsch           #+#    #+#             */
-/*   Updated: 2014/11/12 05:37:50 by jpirsch          ###   ########.fr       */
+/*   Updated: 2014/11/18 19:22:48 by jpirsch          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (i < n)
