+++ 42.1/Libft/ft_strncat.c
@@ -6,7 +6,7 @@
 /*   By: jpirsch <jpirsch@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/11/08 06:51:43 by jpirsch           #+#    #+#             */
-/*   Updated: 2014/11/08 16:04:03 by jpirsch          ###   ########.fr       */
+/*   Updated: 2014/11/18 19:23:38 by jpirsch          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strncat(char *s1, const char *s2, size_t n)
 {
-	int i;
-	int j;
+	size_t i;
+	size_t j;
 
 	i = 0;
 	j = 0;
